Add model tag automatically when calling XXXModel.create({...})
