+++ src/main/dynamo/types.ts
@@ -1,2 +1,4 @@
 import { Model } from "./Model"
+
 export type ExtractFields<T> = T extends Model<infer U, any, any> ? U : never
+export type ModelType = Record<string, any> & { model: string }
